requirements.txt
web3==5.7.0
vyper==0.1.0b17
py-solc-x==0.8.1